streamlit_app/requirements.txt
langchain_openai 
langchain_core
streamlit
langchain_community
langserve
fastapi
uvicorn